pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pubfig"
dynamic = ["version"]
description = "Academic-first publication-quality plotting with Matplotlib for scientific papers"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = [
    "matplotlib",
    "scientific-plotting",
    "publication-figures",
    "academic-figures",
    "data-visualization",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Visualization",
]
dependencies = [
    "matplotlib>=3.8",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
]

[project.urls]
Homepage = "https://github.com/Galaxy-Dawn/pubfig"
Repository = "https://github.com/Galaxy-Dawn/pubfig"
Issues = "https://github.com/Galaxy-Dawn/pubfig/issues"

[project.optional-dependencies]
stats = ["statsmodels>=0.14.0"]
dimreduction = ["scikit-learn>=1.3.0"]
raster = ["pillow>=10"]
all = ["pubfig[stats,dimreduction,raster]"]
dev = ["pytest", "ruff", "mypy"]

[tool.hatch.version]
path = "src/pubfig/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/pubfig"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
